# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Consumer Product Recalls API",
    "description": "Open, read-only API for US consumer product recalls from five agencies (CPSC, FDA, USDA, NHTSA, USCG). No key, GET-only, cursor-paginated. A few things worth knowing up front: a recall's `is_active` flag is `null` (not false) for CPSC and NHTSA, which don't track an open/closed status; `classification` uses each agency's own severity scale and can't be compared across agencies; UPC search matches a whole recall, not an individual product; and search is exact — a typo finds nothing.\n\n**Common lookups:** by product name → `GET /products/search?q=`; by **UPC barcode** → `GET /products/search?upc=`; vehicle or boat by identifier → `GET /products/search?model=` / `?hin=`; a single recall → `GET /recalls/{source}/{recall_id}`.",
    "version": "0.1.1"
  },
  "paths": {
...
    "/recalls": {
      "get": {
        "tags": [
          "recalls"
        ],
        "summary": "List recalls, newest first, with filters and pagination.",
        "description": "Recalls across all five agencies, newest first (by when each was most recently published or updated), with cursor pagination — pass a page's `next_cursor` back as `cursor` for the next page. Filters combine with AND across fields; the categorical ones (`source`, `classification`, `lifecycle_status`, `distribution_scope`, `distribution_state`, `distribution_country`) accept multiple values — repeat the param or comma-separate them (`?source=CPSC,FDA`) to match any of them within that field. Each filter notes its own caveats below. The total count is omitted unless you pass `with_total=true`.",
        "operationId": "list_recalls_recalls_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Source"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Issuing agency; repeat or comma-separate to match any of several.",
              "title": "Source"
            },
            "description": "Issuing agency; repeat or comma-separate to match any of several."
          },
          {
            "name": "classification",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "maxLength": 64
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact match on a source's own classification value; repeat/comma-separate to match any of several.",
              "title": "Classification"
            },
            "description": "Exact match on a source's own classification value; repeat/comma-separate to match any of several."
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "CPSC and NHTSA have no open/closed status, so they're null and match neither true nor false.",
              "title": "Is Active"
            },
            "description": "CPSC and NHTSA have no open/closed status, so they're null and match neither true nor false."
          },
          {
            "name": "published_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive, from the start of that day (UTC).",
              "title": "Published After"
            },
            "description": "Inclusive, from the start of that day (UTC)."
          },
          {
            "name": "published_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive, through the end of that day (UTC).",
              "title": "Published Before"
            },
            "description": "Inclusive, through the end of that day (UTC)."
          },
          {
            "name": "firm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring match on the recall's primary firm name only (not co-recalled firms). Slower than the indexed filters.",
              "title": "Firm"
            },
            "description": "Case-insensitive substring match on the recall's primary firm name only (not co-recalled firms). Slower than the indexed filters."
          },
          {
            "name": "distribution_scope",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DistributionScope"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Distribution scope(s) — Nationwide, Regional, International, or Unspecified; repeat/comma-separate to match any of several.",
              "title": "Distribution Scope"
            },
            "description": "Distribution scope(s) — Nationwide, Regional, International, or Unspecified; repeat/comma-separate to match any of several."
          },
          {
            "name": "lifecycle_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "maxLength": 64
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact match on a source's own status value; repeat/comma-separate to match any of several. CPSC and NHTSA have none, so filtering excludes them.",
              "title": "Lifecycle Status"
            },
            "description": "Exact match on a source's own status value; repeat/comma-separate to match any of several. CPSC and NHTSA have none, so filtering excludes them."
          },
          {
            "name": "announced_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "From the start of that day (UTC), by announcement date; recalls without one are excluded.",
              "title": "Announced After"
            },
            "description": "From the start of that day (UTC), by announcement date; recalls without one are excluded."
          },
          {
            "name": "announced_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Through the end of that day (UTC), by announcement date; recalls without one are excluded.",
              "title": "Announced Before"
            },
            "description": "Through the end of that day (UTC), by announcement date; recalls without one are excluded."
          },
          {
            "name": "source_recall_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact agency recall id; unique only together with `source`.",
              "title": "Source Recall Id"
            },
            "description": "Exact agency recall id; unique only together with `source`."
          },
          {
            "name": "firm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[0-9a-f]{32}$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A firm's id; returns every recall where this firm appears in any role (including co-recalled firms) — unlike `firm`, which is a primary-name substring. Get it from a recall's `firms[].firm_id`.",
              "title": "Firm Id"
            },
            "description": "A firm's id; returns every recall where this firm appears in any role (including co-recalled firms) — unlike `firm`, which is a primary-name substring. Get it from a recall's `firms[].firm_id`."
          },
          {
            "name": "distribution_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 2
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "US state code(s); recalls distributed to any of them (FDA and USDA only). Repeat/comma-separate to match any of several.",
              "title": "Distribution State"
            },
            "description": "US state code(s); recalls distributed to any of them (FDA and USDA only). Repeat/comma-separate to match any of several."
          },
          {
            "name": "distribution_country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 2
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "2-letter country code(s); foreign distribution only (US is excluded by design). Repeat/comma-separate to match any of several.",
              "title": "Distribution Country"
            },
            "description": "2-letter country code(s); foreign distribution only (US is excluded by design). Repeat/comma-separate to match any of several."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size (max 100).",
              "default": 25,
              "title": "Limit"
            },
            "description": "Page size (max 100)."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a prior next_cursor.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a prior next_cursor."
          },
          {
            "name": "with_total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also return the total count (an extra query).",
              "default": false,
              "title": "With Total"
            },
            "description": "Also return the total count (an extra query)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RecallSummary_"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/recalls/{source}/{recall_id}": {
      "get": {
        "tags": [
          "recalls"
        ],
        "summary": "Fetch one recall's full record by agency and recall id.",
        "description": "The full record for one recall, identified by its issuing agency and that agency's own recall id (e.g. `CPSC/24-001`). The agency name is case-insensitive. A couple of field notes: `classification`, `risk_level`, and `lifecycle_status` use each agency's own vocabulary; `is_active` is `null` for CPSC and NHTSA; and `distribution_states` is the agency's free-text prose, separate from the parsed `distribution_state_codes`.",
        "operationId": "get_recall_recalls__source___recall_id__get",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Issuing agency (accepted case-insensitively).",
              "title": "Source"
            },
            "description": "Issuing agency (accepted case-insensitively)."
          },
          {
            "name": "recall_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "description": "The agency-native recall id.",
              "title": "Recall Id"
            },
            "description": "The agency-native recall id."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallDetail"
                }
              }
            }
          },
          "404": {
            "description": "ResourceNotFound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/products/search": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Search recalled products by keyword (full-text) or exact identifier.",
        "description": "Search recalled products — \"is my product recalled?\". Keyword (`q`) is a full-text search over product name, description, recall title, and firm, ranked by relevance (`rank`); whole words and prefixes only, no fuzzy or typo search. `hin` and `model` are exact-match lookups. `upc` matches at the recall level, so each hit carries `upc_is_recall_level: true` — a miss means no recall lists that UPC, not that the product is safe. Supply at least one of `q`, `hin`, `model`, or `upc` (otherwise 422); if several are given, `q` wins, then `hin`/`model`, then `upc`. `source` narrows any of these and accepts multiple values (repeat or comma-separate).",
        "operationId": "search_products_products_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Keywords to search for.",
              "title": "Q"
            },
            "description": "Keywords to search for."
          },
          {
            "name": "hin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact Hull ID (USCG boats).",
              "title": "Hin"
            },
            "description": "Exact Hull ID (USCG boats)."
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact product model.",
              "title": "Model"
            },
            "description": "Exact product model."
          },
          {
            "name": "upc",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "description": "Look up recalled products by 12-digit UPC barcode. Matches at the recall level — UPC data is sparse and comes only from CPSC (about 5% of CPSC recalls; absent for FDA, USDA, NHTSA, and USCG). A miss means no recall lists that UPC, **not** that the product is safe.",
              "examples": [
                "012345678905"
              ],
              "title": "Upc"
            },
            "description": "Look up recalled products by 12-digit UPC barcode. Matches at the recall level — UPC data is sparse and comes only from CPSC (about 5% of CPSC recalls; absent for FDA, USDA, NHTSA, and USCG). A miss means no recall lists that UPC, **not** that the product is safe."
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Source"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional agency filter; repeat/comma-separate to match any of several.",
              "title": "Source"
            },
            "description": "Optional agency filter; repeat/comma-separate to match any of several."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size (max 100).",
              "default": 25,
              "title": "Limit"
            },
            "description": "Page size (max 100)."
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from a prior next_cursor.",
              "title": "Cursor"
            },
            "description": "Opaque cursor from a prior next_cursor."
          },
          {
            "name": "with_total",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Also return the total count (an extra query).",
              "default": false,
              "title": "With Total"
            },
            "description": "Also return the total count (an extra query)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProductSearchHit_"
                }
              }
            }
          },
          "400": {
            "description": "BadCursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "InvalidParameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "UpstreamUnavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DistributionScope": {
        "type": "string",
        "enum": [
          "Nationwide",
          "Regional",
          "Unspecified",
          "International"
        ],
        "title": "DistributionScope",
        "description": "Closed 4-value gold enum (dbt accepted_values; 100% NOT NULL). Free 422 on a bad value."
      },
      "ErrorDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "examples": [
              "not_found"
            ]
          },
          "detail": {
            "title": "Detail",
            "examples": [
              "No recall found for CPSC/24-001."
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "examples": [
              "b1d9c6f2-3a1e-4c7e-9f0a-7d2c1e5b8a40"
            ]
          }
        },
        "type": "object",
        "required": [
          "type",
          "detail",
          "request_id"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "Page_ProductSearchHit_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductSearchHit"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "examples": [
              "eyJ2IjpbIjIwMjYtMDEtMDEiLCJhYmMiXX0"
            ]
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "examples": [
              25
            ]
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total",
            "description": "Only when with_total=true.",
            "examples": [
              null
            ]
          }
        },
        "type": "object",
        "required": [
          "items",
          "limit"
        ],
        "title": "Page[ProductSearchHit]"
      },
...
      "Page_RecallSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecallSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "examples": [
              "eyJ2IjpbIjIwMjYtMDEtMDEiLCJhYmMiXX0"
            ]
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "examples": [
              25
            ]
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total",
            "description": "Only when with_total=true.",
            "examples": [
              null
            ]
          }
        },
        "type": "object",
        "required": [
          "items",
          "limit"
        ],
        "title": "Page[RecallSummary]"
      },
...
      "Source": {
        "type": "string",
        "enum": [
          "CPSC",
          "FDA",
          "USDA",
          "NHTSA",
          "USCG"
        ],
        "title": "Source",
        "description": "The one closed cross-source domain (uppercase). Free 422 on a bad path/query value."
      },
...
    }
  }
}